Set local signaling address when connecting the channel
This CL updates the Corp messaging stack to allow the class which
manages the messaging channel (CorpMessageChannelStrategy) to provide
an updated signaling address when it is set or changes. This is needed
because unlike FTL, the Corp machine may not know the UUID which is
required to build the signaling address so it is provided in the
ChannelOpen message. While it *could* change between channel
open/close cycles, it is currently stable. The code and tests handle
this condition though. I've also addressed a comment from a previous
CL for a string comparison modernization.

Bug: 359620500
Change-Id: I6362194db39e35dfbcd06ca846932bc132fb586d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7268896
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560242}