[Projects] Introducing Feature Flag + Creating View Container Overlay
This change adds the feature flag for the projects panel view as well as
its basic components. The ProjectsPanelStateController acts as a Browser
Window Feature, and the ProjectsPanelView is painted on top of the
VerticalTabStripRegion as a layer.

Some future considerations are that we should take a look at input types
to the project panel overlay (HTClient?) and modify to use FlexLayout.

Change-Id: If4d426bbdd945b4e9c653837bc25a6fcdd4d159f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7249194
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Auto-Submit: Kunal Daftari <kunaldaftari@google.com>
Cr-Commit-Position: refs/heads/main@{#1560251}