[iOS][CX] Repurpose invalid password view for passkeys
Adjusts the invalid password view to allow displaying it for passkeys as
well. The view is meant to display one credential type only (i.e. it
will be displayed after tapping on an info button of the respective
credential row in the main import view controller.

Hence, the initializer is changed to accept CredentialImportItem and
CredentialType.

Bug: 450982128
Change-Id: Iadb92096b6e7bd2fc6749e5984d64469766ab091
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7262374
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1560267}