Update username handling for Corp Signaling
This CL simplifies how usernames are retrieved and handled for
Corp signaling. We now use GetUsername() to retrieve the username
which this host is running as, this prevents issues like running a
host as a different user (per the config0 and then failing due to
the session policy which requires the local username to match the
remote username since the Corp Signaling service will not allow a
different user to send messages to the host.

Along similar lines, there is no need to compare the 'username' of
the incoming signaling message since mismatched values will be
rejected server-side.

Lastly, I had originally intended to gate Corp Signaling behind
having the 'is_corp_host' hint in the config but I think that
will interfere with rolling these changes out since older hosts
will not have that line in their config.

Bug: 359620500
Change-Id: Ida286d193100491cd8610734d209a9aa2f9d1a4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7271607
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560269}