[Contextual Tasks] Add `source=` param to Lens queries
This CL adds the invocation source to the URL for Lens queries that
route to Contextual Tasks. Previously, the source param was not included
since it was not needed for the existing AIM flows. The attached bug
asks to use a different string, but that will come in a follow up to
ensure this does not regress metrics.

Bug: b:467397501
Change-Id: Ibf3ee3f423db94b8d1ee5073368363d2d2e3ba80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7261787
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1560270}