[meta text-scale] Support for Desktop
For Desktop, the meta tag causes us to expose the UA-level font scale to
env(preferred-text-scale). It also will eventually cause Windows to not
do full-browser zoom, but that is not done yet.

This CL makes env() return the product of UA-level and OS-level factors,
but on Desktop the OS-level setting is still always 1, so it effectively
exposes only the UA-level factor.

We clamp the OS-level font setting before exposing it via env() and
applying it to 'medium', but we don't clamp the UA-level font setting.
The UA-level setting is trivially exposed today, even without <meta>.

Bug: 430566925
Change-Id: I77d59457765857c0a03defbf2bcfe33c4065da5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7264325
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560279}