[Settings Search] Let search engine/notification pref open from search results
Search engine settings is based on ListFragment that cannot be casted
to PreferenceFragmentCompat. This CL loosens the type check so that
other types of Fragment may be processed as well. The difference will
be that only the preference in PreferenceFragmentCompat may be
scrolled to and highlighted, since general Fragment is not guaranteed
to have a list view inside.

Notification launches an external activity. It was also made to work
in the same way when opened from search results.

Bug: 469676538
Change-Id: I254b43ac8d577ca3a101055998601298950bc30d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7269247
Reviewed-by: Moe Adel <adelm@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560337}