extension: UninstallMultipleExtensions test
On Android, we do not support the dialog delegate thus we're not able to
get pass the dialog triggered by the remove extension action. However,
there is an `accept_bubble_for_testing` argument which should be a
workaround for testing. This CL replace the window dialog logic with the
`accecpt_bubble_for_testing` argument.

Bug: 439447971
Change-Id: Ic123d2e5b6293c346bb6d87dd93ae4e0c18789ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7267372
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Kay Lin <kaiyilin@google.com>
Cr-Commit-Position: refs/heads/main@{#1560338}