Use SimpleURLLoader::DownloadHeadersOnly in components/feedback
This change replaces calls to SimpleURLLoader's
DownloadToStringOfUnboundedSizeUntilCrashAndDie with DownloadHeadersOnly
in various parts of the Chrome codebase. The body of the response was
not being used in these instances, so having SimpleURLLoader not store
the response body is more efficient and better conveys how the callers
interact with the request responses.

No major changes in functionality are expected as a result of this other
than SimpleURLLoader no longer saving the response bodies while the
request is being downloaded.

This CL was uploaded by git cl split.

Bug: None
Change-Id: I5a01157d14d208ebabcd7fdf7895294fe261febe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270845
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560362}