views: Optimize the animation frame rate of CompositorAnimationRunner.
The CompositorAnimationRunner is driven by the CC's OnBeginFrame method.
CC already limits the frequency of OnBeginFrame, so we don't need to
limit it in the CompositorAnimationRunner. On the contrary, limiting it
in the CompositorAnimationRunner would cause dropped frames in the
animation.

This is because, under CC scheduling, there might be reasons why the
`timestamp - last_tick_ < min_interval_` condition might not always be
met.

Using `--ui-show-fps-counter`, you can see a significant improvement in
the frame rate of adding and deleting tab animations.

Local testing showed a limit of 50fps before the modification, and a
limit of 60fps after the modification (on a 60fps monitor).

Bug: None
Change-Id: I3fddd6e78ba6dea5e25542bc5945c6b6210c9447
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7273144
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Weidong Liu <weidongliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560388}