[UNO] Always show sign in promos in signin pending state
This change is entirely behind the flag `UnoPhase2FollowUp`.

For accounts in signin pending state this shows the bookmark signin
promos even if the user was not yet syncing that data type. For
extensions, the promo was already shown with
https://chromium-review.googlesource.com/c/chromium/src/+/7146763, but
the data type was not enabled upon reauth. This enables the data type
and migrates the data to the account once the promo is accepted.

Note that with this change, the data type (for all sign in promos) is
enabled immediately upon accepting the promo, rather than waiting for
reauth to complete.

There is a bug that removes the newly installed extension. It appears
that this happens only if the extension was previously removed with sync
enabled and is not limited to the promo, see
https://crbug.com/467164248.

Bug: 460403839
Change-Id: Ief9295b83909f7a459c014a17b8e4a6e3f648fea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7242163
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: Amelie Schneider <amelies@google.com>
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560472}