Decouple ad-hoc signature verification from IsolatedWebAppResponseReader
The response reader is a heavy class tailored for fetching resources
from a bundle and converting them to http responses, whereas ad-hoc
signature verification doesn't care about this functionality. For this
reason it's more preferred to interface with the SignedWebBundleReader
directly when it comes to explicit signature verification.

Bug: None
Change-Id: Ib282413cf02d16d1329991c61273ba8fc26e3ac4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7274378
Reviewed-by: Simon Hangl <simonha@google.com>
Commit-Queue: Andrew Rayskiy <greengrape@google.com>
Cr-Commit-Position: refs/heads/main@{#1560490}