manifest: Extend display_override to accept url_patterns
Currently display_override in the web manifest accepts a list of
DisplayMode strings.

This CL extends it to additionally accept objects in the format

  { "display": "unframed", "url_patterns": [ ... ] }

The list of url_patterns can be used to select in which URLs this
display override applies.

manifest_parser.cc permits out-of-scope and out-of-origin URL patterns.
Downstream consumers of these patterns should validate and only apply
the display mode when in-scope.

For the time being only the "unframed" display mode is allowed to have
url_patterns.

Bug: 465492865
Test: testing/xvfb.py tools/autotest.py -C out_/Default third_party/blink/renderer/modules/manifest/manifest_parser_unittest.cc
Change-Id: I9c99ff33b0ff63e7c7c33ae9d32c07c701aeabe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7228926
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Commit-Queue: Edman Anjos <edman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560517}