Use SimpleURLLoader::DownloadHeadersOnly in chrome/browser/profile*
This change replaces calls to SimpleURLLoader's
DownloadToStringOfUnboundedSizeUntilCrashAndDie with DownloadHeadersOnly
in various parts of the Chrome codebase. The body of the response was
not being used in these instances, so having SimpleURLLoader not store
the response body is more efficient and better conveys how the callers
interact with the request responses.

No major changes in functionality are expected as a result of this other
than SimpleURLLoader no longer saving the response bodies while the
request is being downloaded.

This CL was uploaded by git cl split.

Bug: None
Change-Id: I8afaa3dd652c512ac4d5dde7e973af3ec4257eae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270427
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560551}