[Android] Implement undocked Keyboard Accessory Fallback Menu
This change introduces undocked styling for the accessory sheet on
large-screen Android devices. Instead of being docked at the bottom of
the screen, the sheet will now appear as a floating pane with a fixed
maximum width.

The positioning logic is triggered in ManualFillingMediator, which
determines if the undocked style should be applied based on the device
form factor. The AccessorySheetMediator then calculates the specific
layout properties, utilizing the BrowserControlsManager to determine the
correct top offset relative to the content area.

This new positioning logic is passed down via the property model, which
then uses the values to apply the correct styling to the
AccessorySheetView.

The change is hidden behind the
AUTOFILL_ANDROID_KEYBOARD_ACCESSORY_DYNAMIC_POSITIONING feature flag.

Correct height handling and further style changes will be implemented in
the following CLs.

Demo: http://screencast/cast/NDg4ODEyNDcxODA1NTQyNHw5ODY4NmExNC03Yg
Bug: 458654729
Change-Id: Ia85b074118b5fc249a6414a1dd5c754e16609ff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7264889
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Piotr Kotynia <piotrkotynia@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Reviewed-by: Luchen Peng <luchenpeng@google.com>
Cr-Commit-Position: refs/heads/main@{#1560575}