[iOS][CX] Add correct banners in credential import
Adds displaying a correct banner in the import flow based on the set
exported display name. For now, adds a banner for Apple Passwords app
and a generic one, that should be used when e.g. exporter display name
is not set or does not match any known ones.

https://screenshot.googleplex.com/8kAsQ3vYbRoCesQ
https://screenshot.googleplex.com/9uqFxMTHvWyN9Pz

Bug: 450982128
Change-Id: I45ccef7a7395cef3b25cb001ba3dcf5bb2d1ebf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270084
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1560604}