[LNA] Add ContentSetting and PageInfo for LNA Split Permissions
This adds UI for controlling the split LNA permissions and content settings in Page Info and Site Setting on Desktop and Android.

Content settings UIs are gated by the feature flag; page info is not as it should not show if the content settings is on a default value, which it will be unless its modified by the UI (which is feature flag gated).

Strings are based on approved strings in go/split-lna-permission-dd; loopback network icons for desktop and android are placeholders and will be updated in a later CL, as will adding the omnibox chip.

Screenshots:

Content settings desktop: https://crbug.com/465491626#comment4
Content settings android: https://crbug.com/465491626#comment5
Page Info(both): https://crbug.com/465491626#comment6

CLs for adding the new permissions:

1. https://crrev.com/c/7225793/: permissions/content settings types
2. this CL: content settings and page info UI
3. https://crrev.com/c/7243192/: permission prompt UI
4. https://crrev.com/c/7260483/: permission policies

Bug: 465491626
Change-Id: Idd86c297ffb4de8b3edadabd334fb976bbb68f8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7233078
Reviewed-by: Judith Hemp <hempjudith@google.com>
Commit-Queue: Hubert Chao <hchao@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Nicola Tommasi <tommasin@chromium.org>
Reviewed-by: Yifan Luo <lyf@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560617}