[dbsc] Add UnexportableKeyProfileGarbageCollectionServiceFactory
This change adds `UnexportableKeyProfileGarbageCollectionServiceFactory`
to manage the lifecycle of unexportable keys. For off-the-record
profiles, a service is created to delete all associated keys upon
profile shutdown. For regular profiles, a delayed garbage collection
task is scheduled to identify and delete keys that were created for
now-obsolete off-the-record profiles. New tests are added to verify the
garbage collection logic on macOS.

Bug: 455538832
Change-Id: Icfae1d09304f14f861575a946a71b21e9aa24626
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7245410
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560624}