[Extension] Include extension root in content verifier cache key.
Before this change we didn't consider the extension root when creating
content verification cache keys, but, as we've seen in crashes, there
can be mismatches on the extension root when a content verification job
is created until it is started. This can occur during corruption repair
where a new versioned directory (e.g., with a _N suffix) is created (if
the repair must wait for the extension idle).

After this change we include the extension root directory in the
ContentHash cache key to prevent mismatching hashes when an extension
(temporarily, until garbage collection) has multiple root directories
for the same version. By including the root path in the cache key, we
ensure that verification jobs use the hashes exactly matching the
extension root they were created for.

I'd love to create an e2e browsertest for this, but the timing seems a
bit too tricky to feasibly create one. The unit test included verifies
the basic functionality added here though.

Bug: 416484593
Change-Id: Ifb2ccfd6032de09b4a72fc6d1de16b4e0722df07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266825
Reviewed-by: Andrea Orru <andreaorru@chromium.org>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560678}