[context-input] Add CHECK for enterprise policy checking
Adds a method to the session handle that checks for the enterprise
policy. If this method is not called before context is uploaded, this
will trigger a crash.

Also adds calls to this new method wherever SessionHandles are created.

Tested locally with omnibox popup flow, ntp flow, and manual navigation
to contextual-tasks, and issuing context uploads, without crashes.

No changes are need on Android because the query controller bridge is
used, bypassing the session handle.

Bug: b:469148340, b:469875164, b:469877869, b:469875837, b:469877646, b:469875247, b:469875271, b:470072061
Change-Id: I90d6dba87914ca08790649d44d150177e2cb9660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270996
Reviewed-by: Christian Xu <christianxu@chromium.org>
Commit-Queue: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1560682}