[PWA] Suppress fatal asserts in BTMDatabase from SQLITE I/O errors
The BTMDatabase is initialized for different profiles, and deleting
them in the middle of testing fails with a
`SQLITE_IOERR_DELETE_NOENT` error while running SQLITE queries as
part of initialization, as a result of calling
sql::Database::IsExpectedSqliteError().

Since crrev.com/c/7264555 has gone in, the guidance on the exact
logging mechanisms for these errors have changed, so this CL
updates the BTMDatabase to use DLOG(ERROR)s instead of fatal
assertions.

As a consequence, this fixes flakiness in ALL profile deletion
tests in Chrome which fail with I/O errors in the BTMDatabase
initialization, with
WebAppProfileDeletionTest_WebContentsGracefulShutdown being one of
them.

Fixed: 454830629
Change-Id: I7f42d39915166eb68821035c3aeda1a9c8809eea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7263684
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Joshua Hood <jdh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560691}